[#54] wip: :gear: 백엔드 타입스크립트 확장자 변경
